backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.4
numpy==1.26.3
whitenoise==6.6.0
gunicorn==21.2.0
